Refactor dispo history tags into separate component
