+++ frontend/src/components/Observables/ObservableLeaf.vue
@@ -86,15 +86,10 @@
       v-if="showDispositionTags && observable.dispositionHistory.length"
       class="leaf-element"
     >
-      <AlertDispositionTag
-        v-for="entry in observable.dispositionHistory"
-        :key="entry.disposition"
-        style="cursor: pointer"
-        :disposition="entry.disposition"
-        :disposition-count="entry.count"
-        :percent="entry.percent"
-        @click="filterByObservableAndDisposition(observable, entry.disposition)"
-      ></AlertDispositionTag>
+      <ObservableDispositionHistoryGroup
+        :observable="observable"
+        :reroute-to-manage-alerts="true"
+      ></ObservableDispositionHistoryGroup>
     </span>
   </span>
 </template>
@@ -119,7 +114,7 @@
 
   import type CSS from "csstype";
 
-  import AlertDispositionTag from "@/components/Alerts/AlertDispositionTag.vue";
+  import ObservableDispositionHistoryGroup from "@/components/Observables/ObservableDispositionHistoryGroup.vue";
   import MetadataTag from "@/components/Metadata/MetadataTag.vue";
 
   import {
@@ -132,10 +127,8 @@
   import type { observableActionSubTypes } from "@/models/observable";
   import { copyToClipboard, prettyPrintDateTime } from "@/etc/helpers";
   import { useAlertStore } from "@/stores/alert";
-  import { useAlertDispositionStore } from "@/stores/alertDisposition";
   import { useFilterStore } from "@/stores/filter";
   import { useModalStore } from "@/stores/modal";
-  import { alertDispositionRead } from "@/models/alertDisposition";
 
   const config = inject("config") as Record<string, any>;
 
@@ -153,7 +146,6 @@
   });
 
   const alertStore = useAlertStore();
-  const alertDispositionStore = useAlertDispositionStore();
   const filterStore = useFilterStore();
   const modalStore = useModalStore();
   const toast = useToast();
@@ -292,47 +284,6 @@
     router.replace({
       path: "/manage_alerts",
     });
-  };
-
-  const filterByObservableAndDisposition = (
-    obs: observableTreeRead,
-    disposition: string,
-  ) => {
-    const dispositionObject = getDispositionObject(disposition);
-
-    filterStore.clearAll({ nodeType: "alerts" });
-
-    if (dispositionObject) {
-      filterStore.setFilter({
-        nodeType: "alerts",
-        filterName: "disposition",
-        filterValue: dispositionObject,
-      });
-    }
-
-    filterStore.setFilter({
-      nodeType: "alerts",
-      filterName: "observable",
-      filterValue: {
-        category: obs.type,
-        value: obs.value,
-      },
-    });
-    router.replace({
-      path: "/manage_alerts",
-    });
-  };
-
-  const getDispositionObject = (
-    disposition: string,
-  ): alertDispositionRead | undefined => {
-    if (disposition == "OPEN") {
-      return { value: "None" } as alertDispositionRead;
-    } else {
-      return alertDispositionStore.items.find(
-        (item) => item.value == disposition,
-      );
-    }
   };
 
   const showError = (args: {
